requirements.txt
python-hcl2==4.3.5
networkx==3.2.1
colorama==0.4.6
